% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Either.R
\name{Either}
\alias{Either}
\title{'Either' objects}
\description{
R6 class to create and manipulate 'Either' values. An 'Either'
  value is given by a side (the container), 'Left' or 'Right', and an
  element assigned to this side.
}
\note{
Alternatively, you can use the functions \code{\link{Left}} or
  \code{\link{Right}}.
}
\examples{

## ------------------------------------------------
## Method `Either$new`
## ------------------------------------------------

Either$new("left", 123)


## ------------------------------------------------
## Method `Either$isLeft`
## ------------------------------------------------

Either$new("right", 999)$isLeft()

## ------------------------------------------------
## Method `Either$isRight`
## ------------------------------------------------

Either$new("right", 999)$isRight()

## ------------------------------------------------
## Method `Either$getLeft`
## ------------------------------------------------

Either$new("right", 999)$getLeft("abc")

## ------------------------------------------------
## Method `Either$getRight`
## ------------------------------------------------

Either$new("right", 999)$getRight("abc")

## ------------------------------------------------
## Method `Either$toMaybe`
## ------------------------------------------------

Either$new("right", 999)$toMaybe()
Either$new("left", 999)$toMaybe()

## ------------------------------------------------
## Method `Either$mapLeft`
## ------------------------------------------------

Either$new("left", 999)$mapLeft(is.numeric)
Either$new("right", 999)$mapLeft(is.numeric)

## ------------------------------------------------
## Method `Either$mapRight`
## ------------------------------------------------

Either$new("left", 999)$mapRight(is.numeric)
Either$new("right", 999)$mapRight(is.numeric)

## ------------------------------------------------
## Method `Either$either`
## ------------------------------------------------

Either$new("left", 999)$either(is.numeric, is.character)
Either$new("right", 999)$either(is.numeric, is.character)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Either$new()}}
\item \href{#method-print}{\code{Either$print()}}
\item \href{#method-isLeft}{\code{Either$isLeft()}}
\item \href{#method-isRight}{\code{Either$isRight()}}
\item \href{#method-getLeft}{\code{Either$getLeft()}}
\item \href{#method-getRight}{\code{Either$getRight()}}
\item \href{#method-toMaybe}{\code{Either$toMaybe()}}
\item \href{#method-mapLeft}{\code{Either$mapLeft()}}
\item \href{#method-mapRight}{\code{Either$mapRight()}}
\item \href{#method-either}{\code{Either$either()}}
\item \href{#method-clone}{\code{Either$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new 'Either' object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Either$new(side, element)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{side}}{the container, \code{"Left"} or \code{"Right"} (lower case
is accepted)}

\item{\code{element}}{the element to wrap in the container}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An 'Either' object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Either$new("left", 123)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints an 'Either' object (this method is automatically
  called).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Either$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isLeft"></a>}}
\if{latex}{\out{\hypertarget{method-isLeft}{}}}
\subsection{Method \code{isLeft()}}{
Checks whether the reference 'Either' object is a 'Left'
  value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Either$isLeft()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A Boolean value.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Either$new("right", 999)$isLeft()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isRight"></a>}}
\if{latex}{\out{\hypertarget{method-isRight}{}}}
\subsection{Method \code{isRight()}}{
Checks whether the reference 'Either' object is a 'Right'
  value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Either$isRight()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A Boolean value.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Either$new("right", 999)$isRight()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLeft"></a>}}
\if{latex}{\out{\hypertarget{method-getLeft}{}}}
\subsection{Method \code{getLeft()}}{
Get the content of a 'Left' value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Either$getLeft(default)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{default}}{a default element to return in case if the reference
'Either' object is not a 'Left' value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The content of the reference 'Either' object if it is a 'Left'
  value, otherwise the default element.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Either$new("right", 999)$getLeft("abc")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRight"></a>}}
\if{latex}{\out{\hypertarget{method-getRight}{}}}
\subsection{Method \code{getRight()}}{
Get the content of a 'Right' value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Either$getRight(default)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{default}}{a default element to return in case if the reference
'Either' object is not a 'Right' value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The content of the reference 'Either' object if it is a 'Right'
  value, otherwise the default element.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Either$new("right", 999)$getRight("abc")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toMaybe"></a>}}
\if{latex}{\out{\hypertarget{method-toMaybe}{}}}
\subsection{Method \code{toMaybe()}}{
Converts an 'Either' object to a 'maybe' object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Either$toMaybe()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The 'nothing' value if the reference 'Either' object is a 'Left'
  value, otherwise the 'Just' value with the same content as the
  reference 'Either' object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Either$new("right", 999)$toMaybe()
Either$new("left", 999)$toMaybe()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapLeft"></a>}}
\if{latex}{\out{\hypertarget{method-mapLeft}{}}}
\subsection{Method \code{mapLeft()}}{
Apply a function to the content of the reference 'Either'
  object if it is a 'Left' value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Either$mapLeft(f)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{a function}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If the reference 'Either' object is a 'Left' value, this returns
  the 'Left' value whose content is the result of \code{f} applied to
  the original content. Otherwise this returns a clone of the reference
  'Either' object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Either$new("left", 999)$mapLeft(is.numeric)
Either$new("right", 999)$mapLeft(is.numeric)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapRight"></a>}}
\if{latex}{\out{\hypertarget{method-mapRight}{}}}
\subsection{Method \code{mapRight()}}{
Apply a function to the content of the reference 'Either'
  object if it is a 'Right' value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Either$mapRight(f)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{a function}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If the reference 'Either' object is a 'Right' value, this returns
  the 'Right' value whose content is the result of \code{f} applied to
  the original content. Otherwise this returns a clone of the reference
  'Either' object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Either$new("left", 999)$mapRight(is.numeric)
Either$new("right", 999)$mapRight(is.numeric)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-either"></a>}}
\if{latex}{\out{\hypertarget{method-either}{}}}
\subsection{Method \code{either()}}{
Apply a function to the content of the reference 'Either'
  object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Either$either(f, g)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{function to apply to the content of the reference 'Either'
object if it is a 'Left' value}

\item{\code{g}}{function to apply to the content of the reference 'Either'
object if it is a 'Right' value}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An 'Either' object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Either$new("left", 999)$either(is.numeric, is.character)
Either$new("right", 999)$either(is.numeric, is.character)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Either$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
