% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Either.R
\name{eitherFromMaybe}
\alias{eitherFromMaybe}
\title{'Either' object from 'maybe' object}
\usage{
eitherFromMaybe(a, mb)
}
\arguments{
\item{a}{content of the potential 'Left' output value}

\item{mb}{a 'maybe' value; if \code{nothing}, the function returns the
'Left' value containing \code{a}, otherwise the function returns the
'Right' value containing the content of the 'Just' \code{mb} value}
}
\value{
An \code{\link{Either}} object.
}
\description{
Creates a new \code{\link{Either}} object from a 'maybe' object.
}
\examples{
library(maybe)
eitherFromMaybe("abc", nothing())
eitherFromMaybe("abc", just(123))
}
