% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Either.R
\name{Left}
\alias{Left}
\title{New 'Left' value}
\usage{
Left(element)
}
\arguments{
\item{element}{the content of the 'Left' value}
}
\value{
A 'Left' \code{\link{Either}} object.
}
\description{
Creates a new \code{\link{Either}} object of type 'Left'.
}
\examples{
Left("abc")
}
