% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Either.R
\name{Right}
\alias{Right}
\title{New 'Right' value}
\usage{
Right(element)
}
\arguments{
\item{element}{the content of the 'Right' value}
}
\value{
A 'Right' \code{\link{Either}} object.
}
\description{
Creates a new \code{\link{Either}} object of type 'Right'.
}
\examples{
Right("abc")
}
