\name{EITHER}
\alias{EITHER}
\title{EITHER evaluates the tumor heterogeneity level of each tumor sample based on Shannon Entropy Heterogeneity score vignette}
\usage{
EITHER()
}
\description{
EITHER evaluates the tumor heterogeneity level of each tumor sample based on Shannon Entropy Heterogeneity score vignette.
}
\examples{
EITHER(exp, match)
}
